+++ scim2_models/resources/enterprise_user.py
@@ -22,7 +22,7 @@ class Manager(ComplexAttribute):
     ref: Annotated[  # type: ignore[type-arg]
         Reference["User"] | None,
         Required.true,
-    ] = Field(None, serialization_alias="$ref")
+    ] = Field(None, serialization_alias="$ref", validation_alias="$ref")
     """The URI of the SCIM resource representing the User's manager."""
 
     display_name: Annotated[str | None, Mutability.read_only] = None