+++ scim2_models/resources/user.py
@@ -200,7 +200,7 @@ class GroupMembership(ComplexAttribute):
     ref: Annotated[
         Reference["Group"] | None,
         Mutability.read_only,
-    ] = Field(None, serialization_alias="$ref")
+    ] = Field(None, serialization_alias="$ref", validation_alias="$ref")
     """The reference URI of a target resource, if the attribute is a
     reference."""